Move config code to own file
